+++ lib/modules/about/version_history.dart
@@ -1,5 +1,4 @@
 import 'dart:convert';
-
 import 'package:remixicon/remixicon.dart';
 import 'package:pure_live/common/index.dart';
 import 'package:markdown_widget/widget/all.dart';
@@ -44,7 +43,7 @@ class _VersionHistoryPageState extends State<VersionHistoryPage> with SingleTick
     if (historyLoading.value) return;
 
     try {
-      GitHubMirror mirror = GitHubMirror(owner: 'wzgrx', repo: 'pure_live', branch: 'master');
+      GitHubMirror mirror = GitHubMirror(owner: 'liuchuancong', repo: 'pure_live', branch: 'master');
       historyLoading.value = true;
       historyError.value = false;
       final timestamp = DateTime.now().millisecondsSinceEpoch;
@@ -59,7 +58,8 @@ class _VersionHistoryPageState extends State<VersionHistoryPage> with SingleTick
       var result = await HttpClient.instance.getJson(
         url,
         header: {
-          'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/114.0.0.0 Safari/537.36 Edg/114.0.1823.51',
+          'User-Agent':
+              'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/114.0.0.0 Safari/537.36 Edg/114.0.1823.51',
           'Accept': 'application/json',
         },
       );